server/requirements.txt
Flask
kafka-python
psycopg2-binary
pydantic-settings
